Rename ProposalComparator view to ElementsComparation
Fix #251
